move and sort
