+++ facefusion/types.py
@@ -16,7 +16,8 @@ Prediction : TypeAlias = NDArray[Any]
 KalmanMeasurement : TypeAlias = NDArray[Any]
 KalmanMean : TypeAlias = NDArray[Any]
 KalmanCovariance : TypeAlias = NDArray[Any]
-KalmanCostMatrix : TypeAlias = NDArray[Any]
+KalmanMatrix : TypeAlias = NDArray[Any]
+CostMatrix : TypeAlias = NDArray[Any]
 
 BoundingBox : TypeAlias = NDArray[Any]
 FaceLandmark5 : TypeAlias = NDArray[Any]
